Address review: extract resolvedConfigsEqual + fix CI typecheck
Extract the shared resolvedConfigsEqual helper into
vs/sessions/common/agentHostSessionsProvider.ts and import it from both
the local and remote agent host session providers (was duplicated).

Fix a typecheck failure in agentService.test.ts where a real URI was
being passed to AgentHostStateManager.removeSession, which takes the
protocol-level URI alias (a string).

Also clarify in the createSession comment why we persist the full
resolved config values rather than only the user input: clients render
the resolved state on restore without having to re-resolve.

(Written by Copilot)
