+++ src/vs/sessions/contrib/localAgentHost/browser/localAgentHostSessionsProvider.ts
@@ -31,7 +31,7 @@ import { isSessionConfigComplete } from '../../../common/sessionConfig.js';
 import { diffsToChanges, diffsEqual, mapProtocolStatus } from '../../../common/agentHostDiffs.js';
 import { NotificationType } from '../../../../platform/agentHost/common/state/protocol/notifications.js';
 import { ISendRequestOptions, ISessionChangeEvent } from '../../../services/sessions/common/sessionsProvider.js';
-import { IAgentHostSessionsProvider } from '../../../common/agentHostSessionsProvider.js';
+import { IAgentHostSessionsProvider, resolvedConfigsEqual } from '../../../common/agentHostSessionsProvider.js';
 import { IChat, ISession, ISessionWorkspace, ISessionWorkspaceBrowseAction, SessionStatus, type IGitHubInfo, ISessionType } from '../../../services/sessions/common/session.js';
 
 const LOCAL_PROVIDER_ID = 'local-agent-host';
@@ -58,37 +58,6 @@ function buildMutableConfigSchema(config: Record<string, string>): Record<string
 		};
 	}
 	return properties;
-}
-
-/**
- * Shallow structural equality for resolved session configs. Returns true when
- * both values record the same keys with the same string values and the same
- * set of property keys in the schema. The schema property objects themselves
- * are compared by identity since they originate from the same protocol
- * snapshot.
- */
-function resolvedConfigsEqual(a: IResolveSessionConfigResult, b: IResolveSessionConfigResult): boolean {
-	const aValueKeys = Object.keys(a.values);
-	const bValueKeys = Object.keys(b.values);
-	if (aValueKeys.length !== bValueKeys.length) {
-		return false;
-	}
-	for (const key of aValueKeys) {
-		if (a.values[key] !== b.values[key]) {
-			return false;
-		}
-	}
-	const aPropKeys = Object.keys(a.schema.properties);
-	const bPropKeys = Object.keys(b.schema.properties);
-	if (aPropKeys.length !== bPropKeys.length) {
-		return false;
-	}
-	for (const key of aPropKeys) {
-		if (a.schema.properties[key] !== b.schema.properties[key]) {
-			return false;
-		}
-	}
-	return true;
 }
 
 /**
